+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -146,7 +146,6 @@ class BootStrap {
                     monitorStartupDelayMins: 1,
                     monitorRepeatNotifyMins: 60,
                     monitorTimeoutSecs: 15,
-                    runInLocalDevMode: false,
                     writeToMonitorLog: true
                 ],
                 groupName: 'xh.io',